+++ src/bos/operators/utils/clients/bos/base.py
@@ -27,6 +27,7 @@ from requests.exceptions import HTTPError, ConnectionError
 from urllib3.exceptions import MaxRetryError
 
 from bos.operators.utils import PROTOCOL, requests_retry_session
+from bos.common.utils import exc_type_msg
 
 LOGGER = logging.getLogger('bos.operators.utils.clients.bos.base')
 
@@ -42,13 +43,13 @@ def log_call_errors(func):
             result = func(*args, **kwargs)
             return result
         except (ConnectionError, MaxRetryError) as e:
-            LOGGER.error("Unable to connect to BOS: {}".format(e))
+            LOGGER.error("Unable to connect to BOS: %s", exc_type_msg(e))
             raise e
         except HTTPError as e:
-            LOGGER.error("Unexpected response from BOS: {}".format(e))
+            LOGGER.error("Unexpected response from BOS: %s", exc_type_msg(e))
             raise e
         except json.JSONDecodeError as e:
-            LOGGER.error("Non-JSON response from BOS: {}".format(e))
+            LOGGER.error("Non-JSON response from BOS: %s", exc_type_msg(e))
             raise e
 
     return wrap