+++ src/bos/operators/utils/clients/cfs.py
@@ -25,7 +25,7 @@ from collections import defaultdict
 import logging
 from requests.exceptions import HTTPError, ConnectionError
 
-from bos.common.utils import compact_response_text
+from bos.common.utils import compact_response_text, exc_type_msg
 from bos.operators.utils import requests_retry_session, PROTOCOL
 
 SERVICE_NAME = 'cray-cfs-api'
@@ -48,7 +48,7 @@ def get_components(session=None, **kwargs):
     try:
         response.raise_for_status()
     except HTTPError as err:
-        LOGGER.error("Failed getting nodes from cfs: %s", err)
+        LOGGER.error("Failed getting nodes from CFS: %s", exc_type_msg(err))
         raise
     component_list = response.json()
     LOGGER.debug("Returning %d components from CFS", len(component_list))
@@ -68,7 +68,7 @@ def patch_components(data, session=None):
     try:
         response.raise_for_status()
     except HTTPError as err:
-        LOGGER.error("Failed asking CFS to configure nodes: %s", err)
+        LOGGER.error("Failed asking CFS to configure nodes: %s", exc_type_msg(err))
         raise