+++ src/bos/server/controllers/v2/sessiontemplates.py
@@ -24,6 +24,7 @@
 import logging
 import connexion
 
+from bos.common.utils import exc_type_msg
 from bos.server.models.v2_session_template import V2SessionTemplate as SessionTemplate  # noqa: E501
 from bos.server import redis_db_utils as dbutils
 from bos.server.utils import _canonize_xname
@@ -90,6 +91,7 @@ def put_v2_sessiontemplate(session_template_id):  # noqa: E501
     try:
         data = connexion.request.get_json()
     except Exception as err:
+        LOGGER.error("Error parsing request data: %s", exc_type_msg(err))
         return connexion.problem(
             status=400, title="Error parsing the data provided.",
             detail=str(err))
@@ -106,6 +108,7 @@ def put_v2_sessiontemplate(session_template_id):  # noqa: E501
         """
         SessionTemplate.from_dict(template_data)
     except Exception as err:
+        LOGGER.error("Error creating session template: %s", exc_type_msg(err))
         return connexion.problem(
             status=400, title="The session template could not be created.",
             detail=str(err))
@@ -137,6 +140,7 @@ def get_v2_sessiontemplate(session_template_id):
     """
     LOGGER.debug("GET /v2/sessiontemplates/%s invoked get_v2_sessiontemplate", session_template_id)
     if session_template_id not in DB:
+        LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
             status=404, title="Sessiontemplate could not found.",
             detail="Sessiontemplate {} could not be found".format(session_template_id))
@@ -164,6 +168,7 @@ def delete_v2_sessiontemplate(session_template_id):
     """
     LOGGER.debug("DELETE /v2/sessiontemplates/%s invoked delete_v2_sessiontemplate", session_template_id)
     if session_template_id not in DB:
+        LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
             status=404, title="Sessiontemplate could not found.",
             detail="Sessiontemplate {} could not be found".format(session_template_id))
@@ -180,6 +185,7 @@ def patch_v2_sessiontemplate(session_template_id):
     LOGGER.debug("PATCH /v2/sessiontemplates/%s invoked patch_v2_sessiontemplate", session_template_id)
 
     if session_template_id not in DB:
+        LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
             status=404, title="Sessiontemplate could not found.",
             detail="Sessiontemplate {} could not be found".format(session_template_id))
@@ -194,6 +200,7 @@ def patch_v2_sessiontemplate(session_template_id):
     try:
         data = connexion.request.get_json()
     except Exception as err:
+        LOGGER.error("Error parsing request data: %s", exc_type_msg(err))
         return connexion.problem(
             status=400, title="Error parsing the data provided.",
             detail=str(err))
@@ -210,8 +217,9 @@ def patch_v2_sessiontemplate(session_template_id):
         """
         SessionTemplate.from_dict(template_data)
     except Exception as err:
+        LOGGER.error("Error patching session template: %s", exc_type_msg(err))
         return connexion.problem(
-            status=400, title="The session template could not be created.",
+            status=400, title="The session template could not be patched.",
             detail=str(err))
 
     template_data = _sanitize_xnames(template_data)